rename renderers module to io
- swap pandas requirement with geopandas
- update docs
