+++ docs/io.md
@@ -1,0 +1,3 @@
+# Renderers
+
+::: seasnake.io
