+++ seasnake/__init__.py
@@ -1,3 +1,3 @@
 from .auth import MermaidAuth  # noqa: F401
 from .projects import Project  # noqa: F401
-from .renderers import to_geojson  # noqa: F401
+from .io import to_geojson  # noqa: F401
