+++ tests/test_io.py
@@ -2,7 +2,7 @@ import json
 
 import pytest
 
-from seasnake.renderers import to_geojson
+from seasnake.io import to_geojson
 
 def test_to_geojson(geo_dataframe):
     geojson_str = to_geojson(geo_dataframe)
